requirements.txt
streamlit==1.25.0 
tensorflow==2.13.0 
numpy==1.23.5 
Pillow==9.4.0  
opencv-python==4.8.0.76 